+++ src/errors.rs
@@ -1,6 +1,6 @@
 //! Error type for `oersted`  
 //!
-///! WIP: perhaps this should be multiple errors?
+//! WIP: perhaps this should be multiple errors?
 
 #[derive(Debug)]
 pub enum OerstedError {